requirements.txt
streamlit
transformers
torch
deep-translator
python-docx
PyMuPDF
textblob
textstat
langdetect
pandas 